added scheduled task backup
This is an XML that can be used with Task Scheduler to change the app theme between dark to light at specified times. This is meant to be imported into Windows versions 1903 and later. Importing into earlier versions will cause the XMLs to fail (because the flags they attempt to edit will not be available in earlier versions).